correct availableForWrite signature

+++ cores/nRF5/Adafruit_TinyUSB_Core/Adafruit_USBD_CDC.cpp
@@ -97,7 +97,7 @@ int Adafruit_USBD_CDC::peek(void)
 
 int Adafruit_USBD_CDC::read(void)
 {
-  return tud_cdc_read_char();
+  return (int) tud_cdc_read_char();
 }
 
 void Adafruit_USBD_CDC::flush(void)
@@ -126,7 +126,7 @@ size_t Adafruit_USBD_CDC::write(const uint8_t *buffer, size_t size)
   return size - remain;
 }
 
-size_t Adafruit_USBD_CDC::availableForWrite(void)
+int Adafruit_USBD_CDC::availableForWrite(void)
 {
   return tud_cdc_write_available();
 }
